[MOSIP-42697] Added code for making the info cache configurable.
